+++ arena_tactics/network/client/client.gd
@@ -3,8 +3,8 @@ class_name NetworkClient
 
 signal player_disconnected(peer_id)
 signal server_disconnected
-signal player_list_updated(list: Dictionary)
-signal team_list_updated(list: Dictionary)
+signal player_list_updated(list: Array[ConnectedUser])
+signal team_list_updated(list: Array[ConnectedTeam])
 signal allowed_in_server
 
 signal new_error(msg: String)
@@ -14,8 +14,8 @@ const PORT = 7000
 const DEFAULT_SERVER_IP = "127.0.0.1"
 const MAX_CONNEXION = 4
 
-var players = {}
-var teams = {}
+var users: Array[ConnectedUser] = []
+var teams: Array[ConnectedTeam] = []
 
 var player_info: Dictionary = {"name": "Client", "password": ""}
 
@@ -24,7 +24,6 @@ var player_info: Dictionary = {"name": "Client", "password": ""}
 func set_player_info(_player_info: Dictionary):
 	player_info = _player_info
 
-
 func _ready():
 		multiplayer.peer_connected.connect(_on_player_connected)
 		multiplayer.peer_disconnected.connect(_on_player_disconnected)
@@ -32,7 +31,7 @@ func _ready():
 		multiplayer.connection_failed.connect(_on_connected_fail)
 		multiplayer.server_disconnected.connect(_on_server_disconnected)
 
-#region Players
+#region Users
 func join_game(adress: String = DEFAULT_SERVER_IP, _port: int = PORT) -> bool:
 	leave()
 	var peer := ENetMultiplayerPeer.new()
@@ -54,55 +53,63 @@ func join_game(adress: String = DEFAULT_SERVER_IP, _port: int = PORT) -> bool:
 	return true
 	#print_debug("Client connected to %s:%d" % [adress, PORT])
 
-@rpc("authority", "reliable")
-func allowed_in():
-	allowed_in_server.emit()
-
-func leave():
-	players = {}
-	if multiplayer.has_multiplayer_peer():
-		multiplayer.multiplayer_peer = null
-
-@rpc("any_peer", "reliable")
-func register_player(_new_player_info):
-	pass
-
-@rpc("authority", "reliable")
-func get_player_list(_player_list):
-	players = _player_list
-	player_list_updated.emit()
-
-@rpc("any_peer", "reliable")
-func send_player_list():
-	pass
-
-# when you connect with someone (server or another player)
+# when you connect with someone (server or another user)
 func _on_player_connected(_id):
 	pass
 	#print_debug("%d Joined" % _id)
 
-# when someone disconnect from server
-func _on_player_disconnected(id):
-	players.erase(id)
-	player_disconnected.emit(id)
-	send_player_list.rpc_id(1)
+@rpc("any_peer", "reliable")
+func register_user(_user_name: String, _room_password: String):
+	pass
 
 # when you connect to the server
 func _on_connected_ok():
-	register_player.rpc_id(1, player_info)
-	send_player_list.rpc_id(1)
-	send_team_list.rpc_id(1)
+	register_user.rpc_id(1, player_info["name"], player_info["password"])
+
+@rpc("authority", "reliable")
+func allowed_in():
+	allowed_in_server.emit()
+
+@rpc("authority", "reliable")
+func send_user_list(_users: Dictionary):
+	users.clear()
+	for user in _users:
+		var new_user: ConnectedUser = ConnectedUser.new().from_dict(_users[user])
+		users.append(new_user)
+	player_list_updated.emit()
+
+@rpc("any_peer", "reliable")
+func get_user_list():
+	pass
+
+func leave():
+	users = []
+	if multiplayer.has_multiplayer_peer():
+		multiplayer.multiplayer_peer = null
+
+# when someone disconnect from server
+func _on_player_disconnected(id):
+	player_disconnected.emit(id)
 
 func _on_connected_fail():
 	multiplayer.multiplayer_peer = null
 
 func _on_server_disconnected():
 	multiplayer.multiplayer_peer = null
-	players.clear()
+	users.clear()
 	teams.clear()
 	server_disconnected.emit()
 #endregion
 
+#region UserLib
+
+func get_user_by_id(id: int) -> ConnectedUser:
+	for user in users:
+		if user.id == id:
+			return user
+	return null
+
+#endregion
 
 #region Teams
 func create_team(_team_name: String):
@@ -112,24 +119,32 @@ func create_team(_team_name: String):
 func register_team(_team_name: String):
 	pass
 
-func join_team(_team_name: String):
-	if not teams.has(_team_name):
-		throw_error("No team with name %s was found." % _team_name)
-	add_player_to_team.rpc_id(1, _team_name, player_info)
+func join_team(_team: ConnectedTeam):
+	if not teams.has(_team):
+		throw_error("No team with name %s was found." % _team.team_name)
+	add_player_to_team.rpc_id(1, _team.team_name, player_info["name"])
 
 @rpc("any_peer", "reliable")
-func add_player_to_team(_team_name: String, _player_info: Dictionary):
+func add_player_to_team(_team_name: String, _player_name: String):
 	pass
 
 @rpc("authority", "reliable")
-func get_teams_list(_teams_list):
-	teams = _teams_list
+func send_team_list(_teams: Dictionary):
+	teams.clear()
+	for team in _teams:
+		var new_team: ConnectedTeam = ConnectedTeam.new().from_dict(_teams[team])
+		link_team_id_with_users(new_team)
+		teams.append(new_team)
 	team_list_updated.emit()
 
 @rpc("any_peer", "reliable")
-func send_team_list():
+func get_team_list():
 	pass
 
+func link_team_id_with_users(team: ConnectedTeam):
+	team.members.clear()
+	for id in team.members_id:
+		team.members.append(get_user_by_id(id))
 #endregion
 
 @rpc("authority", "reliable")