+++ arena_tactics/ui/screens/network/lobby/teams/lobby_teams.gd
@@ -6,10 +6,10 @@ var networker: NetworkClient = null
 @export var add_team_button_scene: PackedScene = preload("res://ui/screens/network/lobby/teams/AddTeamButton.tscn")
 @export var team_widget_scene: PackedScene = preload("res://ui/screens/network/lobby/teams/TeamWidget.tscn")
 
-signal picked_team(team_name: String)
+signal picked_team(team: ConnectedTeam)
 signal created_team(team_name: String)
 
-func setup(_networker: NetworkClient, _teams: Dictionary):
+func setup(_networker: NetworkClient):
 	networker = _networker
 	networker.team_list_updated.connect(build_grid)
 	build_grid()
@@ -27,7 +27,7 @@ func fill_grid():
 		var team_widget: TeamWidget = team_widget_scene.instantiate()
 		team_widget.want_join_team.connect(_on_team_picked)
 		self.add_child(team_widget)
-		team_widget.build_widget(team, networker.teams[team])
+		team_widget.build_widget(team)
 	create_add_team_button()
 
 func create_add_team_button():
@@ -35,9 +35,9 @@ func create_add_team_button():
 	inst.pressed.connect(_on_team_created)
 	add_child(inst)
 
-func _on_team_picked(team_name: String):
-	picked_team.emit(team_name)
-	networker.join_team(team_name)
+func _on_team_picked(team: ConnectedTeam):
+	picked_team.emit(team)
+	networker.join_team(team)
 
 func _on_team_created(team_name: String):
 	created_team.emit(team_name)